chore(app): move application to own file
